[MIG] website_sale_product_assortment: Migration to 18.0
