+++ website_sale_product_assortment/static/src/js/assortment_list_preview.esm.js
@@ -1,0 +1,42 @@
+import {renderToElement} from "@web/core/utils/render";
+
+var WebsiteSaleProductAssortment = {
+    selector: "#products_grid",
+
+    start: function () {
+        return $.when.apply($, [
+            this._super.apply(this, arguments),
+            this.render_assortments(),
+        ]);
+    },
+    render_assortments: function () {
+        const $products = $(".o_wsale_product_grid_wrapper");
+        const product_dic = {};
+        $products.each(function () {
+            product_dic[this.querySelector("a img").src.split("/")[6]] = this;
+        });
+        const product_ids = Object.keys(product_dic).map(Number);
+        return this._rpc({
+            route: "/sale/get_info_assortment_preview",
+            params: {product_template_ids: product_ids},
+        }).then((product_values) => {
+            for (const product of product_values) {
+                this.render_product_assortment(product_dic[product.id], product);
+            }
+        });
+    },
+    render_product_assortment: function (product_info, product) {
+        const message_unavailable = renderToElement(
+            "website_sale_product_assortment.product_availability",
+            {
+                message_unavailable: product.message_unavailable,
+                product_template_id: product.id,
+            }
+        );
+        $(product_info).find(".product_price").append(message_unavailable);
+
+        $(product_info).find(".fa-shopping-cart").parent().addClass("disabled");
+    },
+};
+
+export default WebsiteSaleProductAssortment;
