+++ website_sale_product_assortment/static/src/xml/website_sale_product_assortment.xml
@@ -9,12 +9,22 @@
             t-attf-id="message_unavailable_#{product_template_id}"
         >
             <i
-                class="fa fa-exclamation-triangle"
+                class="fa fa-exclamation-triangle me-1"
                 role="img"
                 aria-label="Warning"
                 title="Warning"
             />
             <t t-out="message_unavailable" />
+        </div>
+    </t>
+
+    <t t-name="website_sale_product_assortment.assortment_information">
+        <div class="mt16" id="product_full_assortment_description">
+            <div class="container">
+                <div class="row">
+                    <t t-out="assortment_information" />
+                </div>
+            </div>
         </div>
     </t>
 </templates>
