+++ website_sale_product_assortment/tests/test_ui.py
@@ -18,41 +18,41 @@ class TestUI(HttpCase):
         )
 
     def test_01_ui_no_restriction(self):
+        partner_admin = self.env.ref("base.partner_admin")
         self.env["ir.filters"].create(
             {
                 "name": "Test Assortment",
                 "model_id": "product.product",
                 "is_assortment": True,
                 "domain": [("id", "!=", self.product.product_variant_id.id)],
-                "partner_domain": "[('id', '=', %s)]"
-                % self.env.ref("base.partner_admin").id,
+                "partner_domain": f"[('id', '=', {partner_admin.id})]",
             }
         )
         self.start_tour("/shop", "test_assortment_with_no_restriction", login="admin")
 
     def test_02_ui_no_show(self):
+        partner_admin = self.env.ref("base.partner_admin")
         self.env["ir.filters"].create(
             {
                 "name": "Test Assortment",
                 "model_id": "product.product",
                 "is_assortment": True,
                 "domain": [("id", "!=", self.product.product_variant_id.id)],
-                "partner_domain": "[('id', '=', %s)]"
-                % self.env.ref("base.partner_admin").id,
+                "partner_domain": f"[('id', '=', {partner_admin.id})]",
                 "website_availability": "no_show",
             }
         )
         self.start_tour("/shop", "test_assortment_with_no_show", login="admin")
 
     def test_03_ui_no_purchase(self):
+        partner_admin = self.env.ref("base.partner_admin")
         self.env["ir.filters"].create(
             {
                 "name": "Test Assortment",
                 "model_id": "product.product",
                 "is_assortment": True,
                 "domain": [("id", "!=", self.product.product_variant_id.id)],
-                "partner_domain": "[('id', '=', %s)]"
-                % self.env.ref("base.partner_admin").id,
+                "partner_domain": f"[('id', '=', {partner_admin.id})]",
                 "website_availability": "no_purchase",
                 "message_unavailable": "Can't purchase",
                 "assortment_information": """<span name='testing'>
